moved rest to vertex.py, deleted utils.vertices.py
