+++ ratcave/utils/__init__.py
@@ -1,5 +1,4 @@
 from . import gl
-from . import vertices
 from .gl import vec, create_opengl_object, POINTS, TRIANGLES, LINE_LOOP, LINES, get_viewport, Viewport, clear_color
 from .mixins import NameLabelMixin, BindTargetMixin, BindNoTargetMixin, BindingContextMixin
 from .observers import Observable, Observer, IterObservable, AutoRegisterObserver
